requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-dotenv
supabase
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
email-validator
